general class setup before breaking down modify_user_nl and run_case
